chore: move npm to sub folder
